requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
ez_setup==0.9
gunicorn==23.0.0
idna==3.10
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.1
urllib3==2.2.3
wheel==0.45.0
whitenoise==6.7.0
